pre publish
