+++ README.md
@@ -1,2 +1,14 @@
 # porchmark-pretty-reporter
 Reporter with simple charts and other stuff
+
+Could be used as a plugin for porchmark or as a standalone binary
+
+```
+Options:
+  -c  --config [configfile.js(on)?]  path to config
+  -r --report [report.js(on)?]       path to report. Note: porchamrk do not expose `allMetrics`
+                                     field in its jsonReport. Consider using of this package as a
+                                     plugin
+  -v --verbose                       verbose
+  -h, --help                         display help for command
+```
